tools/requirements.txt
kaitaistruct==0.10
arcade==2.6.17
pyyaml==6.0
